third stage commit backend

+++ src/main/java/com/naja7online/api/dto/UserProfileDto.java
@@ -1,3 +1,3 @@
 package com.naja7online.api.dto;
 
-public record UserProfileDto(Integer id, String name, String email, String level, String role, String avatarUrl) {}
+public record UserProfileDto(Long id, String name, String email, String level, String role, String avatarUrl) {}
